// src/app/models/ResumenConcepto.ts
export class ResumenConcepto{
   

    //RECIBO MASIVO

    idEmpresa: number | null = null;
	idSede: number | null = null;
	idCiclo: number | null = null;
	idSucursal: number | null = null;
	idSector: number | null = null;
	anio: string | null = null;
	mes: number | null = null;
	usuarioCreacion: string | null = null;

	
	idConcepto: number | null = null;
    atrazado: number | null = null;
    ventadirecta: number | null = null;
    fraccionamiento: number | null = null;
    incobrable: number | null = null;
    incobrablefracc: number | null = null;
    importe: number | null = null;
    idTipoConcepto: number | null = null;
    descripcion: string | null = null;



    
}